semantic-ui integration instead of bootstrap
